requirements.txt
flask==3.1.1
pymongo==4.13.2
python-dotenv==1.1.1
requests==2.32.4